refactor: billing/ 廃止 → app/fees/models.py に統合 (F-13 billing cleanup)
- app/billing/ を削除 (v9 モデルは既に不在、V10Base の衝突回避も不要になった)
- FeeConfigV10 / FeeTransaction を app/fees/models.py に移動
- V10Base (分離 metadata) を廃止し app.database.Base に統合
- 全インポート先 6ファイルを更新 (api/fees, calculator, scheduled_tasks, seed, alembic/env, detect_db_gaps)
- テスト: Base.metadata.create_all() 1回で全テーブルを作成するよう簡略化
- 52 tests passed
